requirements.txt
weasyprint
jinja2
huggingface_hub
python-dotenv
fastapi 
uvicorn